Improved: Use common helper function to deduplicate code when using with_split_colour transform
