+++ projects/dxt-lossless-transform-bc1/src/transforms/with_split_colour/transform/avx512.rs
@@ -164,7 +164,6 @@ pub(crate) unsafe fn transform_with_split_colour(
 mod tests {
     use super::*;
     use crate::test_prelude::*;
-    use crate::transforms::with_split_colour::untransform::untransform_with_split_colour;
 
     #[rstest]
     fn avx512_transform_roundtrip() {
@@ -172,34 +171,7 @@ mod tests {
             return;
         }
 
-        for num_blocks in 1..=128 {
-            let input = generate_bc1_test_data(num_blocks);
-            let len = input.len();
-            let mut colour0 = vec![0u16; num_blocks];
-            let mut colour1 = vec![0u16; num_blocks];
-            let mut indices = vec![0u32; num_blocks];
-            let mut reconstructed = vec![0u8; len];
-            unsafe {
-                transform_with_split_colour(
-                    input.as_ptr(),
-                    colour0.as_mut_ptr(),
-                    colour1.as_mut_ptr(),
-                    indices.as_mut_ptr(),
-                    num_blocks,
-                );
-                untransform_with_split_colour(
-                    colour0.as_ptr(),
-                    colour1.as_ptr(),
-                    indices.as_ptr(),
-                    reconstructed.as_mut_ptr(),
-                    num_blocks,
-                );
-            }
-            assert_eq!(
-                reconstructed.as_slice(),
-                input.as_slice(),
-                "Mismatch AVX512 roundtrip split-colour",
-            );
-        }
+        // 256 bytes processed per main loop iteration (* 2 / 8 == 64)
+        run_split_colour_transform_roundtrip_test(transform_with_split_colour, 64, "AVX512");
     }
 }
